Task015 fix

+++ src/main/java/com/epam/university/java/core/task015/Point.java
@@ -8,23 +8,23 @@ public interface Point {
      * Get x value of point.
      * @return value
      */
-    int getX();
+    double getX();
 
     /**
      * Get y value of point.
      * @return value
      */
-    int getY();
+    double getY();
 
     /**
      * Set x value of point.
      * @param x value
      */
-    void setX(int x);
+    void setX(double x);
 
     /**
      * Set y value of point.
      * @param y value
      */
-    void setY(int y);
+    void setY(double y);
 }
